Refactor onboarding and database-update windows to remove duplication
Extract shared installer window chrome, navigation helper, progress bar
controller, database parts picker and download progress UI. Unify the
offline part01/part02 selection on the auto-detect behavior.
