+++ SeforimApp/src/jvmMain/kotlin/io/github/kdroidfilter/seforimapp/features/database/update/navigation/DatabaseUpdateNavHost.kt
@@ -9,13 +9,10 @@ import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import androidx.navigation.NavBackStackEntry
-import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
-import io.github.kdroidfilter.seforim.navigation.NavigationAnimations
 import io.github.kdroidfilter.seforimapp.core.presentation.components.AnimatedHorizontalProgressBar
+import io.github.kdroidfilter.seforimapp.core.presentation.navigation.noAnimatedComposable
 import io.github.kdroidfilter.seforimapp.features.database.update.screens.CompletionScreen
 import io.github.kdroidfilter.seforimapp.features.database.update.screens.OfflineUpdateScreen
 import io.github.kdroidfilter.seforimapp.features.database.update.screens.OnlineUpdateScreen
@@ -70,18 +67,5 @@ fun DatabaseUpdateNavHost(
                 )
             }
         }
-    }
-}
-
-inline fun <reified T : DatabaseUpdateDestination> NavGraphBuilder.noAnimatedComposable(
-    noinline content: @Composable (NavBackStackEntry) -> Unit,
-) {
-    composable<T>(
-        enterTransition = { NavigationAnimations.enterTransition(this) },
-        exitTransition = { NavigationAnimations.exitTransition(this) },
-        popEnterTransition = { NavigationAnimations.popEnterTransition(this) },
-        popExitTransition = { NavigationAnimations.popExitTransition(this) },
-    ) {
-        content(it)
     }
 }
